# swagger-auto.json
{
  "swagger": "2.0",
  "info": {
    "title": "MTG Cards Database",
    "description": "<a id='logout' href='http://cse341-mtg-database.herokuapp.com/auth/logout'>Logout</a>",
    "version": "1.0.0"
  },
  "host": "cse341-mtg-database.herokuapp.com",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "JWT": {
      "type": "apiKey",
      "in": "header",
      "name": "auth-token"
    }
  },
  "paths": {
...
    "/creatures/": {
      "get": {
        "tags": [
          "Creatures"
        ],
        "summary": "Get all creatures",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "Creatures"
        ],
        "summary": "Add new creature to database",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "creatureId": {
                  "example": "any"
                },
                "creatureName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/creatures/{creatureName}": {
      "get": {
        "tags": [
          "Creatures"
        ],
        "summary": "Get one creature by creature name",
        "description": "",
        "parameters": [
          {
            "name": "creatureName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "tags": [
          "Creatures"
        ],
        "summary": "Update existing creature in database",
        "description": "",
        "parameters": [
          {
            "name": "creatureName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "creatureId": {
                  "example": "any"
                },
                "creatureName": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Creatures"
        ],
        "summary": "Delete existing creature from database",
        "description": "",
        "parameters": [
          {
            "name": "creatureName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/abilities/": {
      "get": {
        "tags": [
          "Abilities"
        ],
        "summary": "Get all card abilities",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "Abilities"
        ],
        "summary": "Add new card ability to the database",
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "abilityId": {
                  "example": "any"
                },
                "abilityName": {
                  "example": "any"
                },
                "abilityDescription": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/abilities/{abilityName}": {
      "get": {
        "tags": [
          "Abilities"
        ],
        "summary": "Get one card ability by ability name",
        "description": "",
        "parameters": [
          {
            "name": "abilityName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "put": {
        "tags": [
          "Abilities"
        ],
        "summary": "Update existing card in the database",
        "description": "",
        "parameters": [
          {
            "name": "abilityName",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "abilityId": {
                  "example": "any"
                },
                "abilityName": {
                  "example": "any"
                },
                "abilityDescription": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Abilities"
        ],
        "summary": "Delete existing card from the database",
        "description": "",
        "parameters": [
          {
            "name": "abilityName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    }
  }
}